Multi-layer support. Layers controlled by layers.conf (currently built-in: default layer and maplint layer). Empty layers.conf emulates old behaviour. All layers rendered off the same downloaded data.
Memory consumption of tilesGen.pl may have increased for complex tiles as
maplint demanded a more fine-tuned approach to merging of multiple stripes.

Version name bumped up to "Essen" even though there's no visible changes on the
default layer - I felt the large number of changes merited that.

TODOs: Make community decisions on what maplint tests should be run and how
they are supposed to be rendered. This version uses a rather strict set
of tests. Tests are encoded in tests.xsl which is generated by a Makefile
from within the original maplint distribution; the Makefile and the components
from which tests.xsl is assembled are not included in tiles@home.