Introducing four new icons for parking:
A "P" with a roof, used for multi-storey and underground parking.
A "P" with cash next to it for parking amenities where you have to pay
A "P" with a roof and cash next to it for the combined case.
Those icons are based on work by Wabba

And a "P" with a  car+bus box next to it for park and ride.
This icon is based on work by Batchoy