Make sure that osm2pgsql does not attempt to append data to a table when it already has data in a different SRS.
Without this patch it is perfectly possible for the mindless user (Y.T.) to create a table with -l and later append to it without -l, which will land you with a "select distinct srid(way) from planet_osm_point" returning two SRSs. Mapnik's queries will then fail with an "Operation on two geometries with different SRIDs" error. 

Note, this patch only checks the default SRID given in the geometry_columns table.