An initial (incomplete) "proof of concept" integration of an OpenStreetBugs interface into the rails port
This is a (start of a) reimplementation of the api at http://github.com/emka/openstreetbugs/tree/master/api/0.1/
into rails

The client side code is taken from http://wiki.openstreetmap.org/wiki/OpenStreetBugs/New_Client

This is ment as a mockup to perhaps use as a basis to further discuss how best to integrate a map bug reporting system

It currently uses (more or less) the openstreetbugs api specification. But this api feels rather inconsistent with the
rest of the rails_port api, so depending on discussions might still need some significant changes.